requirements.txt
streamlit
PyYAML
streamlit_authenticator
openai
plotly
streamlit_option_menu
